RDCC-3333: adding null check to request and updating unit tests
